+++ app/decorators/application_decorator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ApplicationDecorator < Draper::Decorator
+  include ActionView::Helpers::TranslationHelper
+
+  delegate_all
+end