+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.3.0] - 2022-02-28
+
+### Features
+- Project contract summary pdf generation endpoint.
+- New gems used for pdf generation:
+    - `wicked_pdf`
+    - `wkhtmltopdf-binary`
+    - `draper`
+
 ## [2.1.1] - 2022-03-09
 
 ### Features